backend/requirements.txt
# Core dependencies
fastapi==0.115.2
uvicorn[standard]==0.31.0
python-multipart==0.0.12
python-dotenv==1.0.1

# Vector database
qdrant-client==1.12.1

# AI/ML
openai==1.53.0
tiktoken==0.8.0

# File processing
pypdf==5.1.0
pypdf2==3.0.1
Pillow==11.0.0
python-magic==0.4.27
python-docx==1.1.0
chardet==5.2.0

# Data processing
numpy<2.0.0  # Use numpy 1.x for compatibility
scikit-learn==1.5.2
pandas==2.2.3
umap-learn==0.5.5

# Testing
pytest==8.3.3
httpx==0.27.2

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.1

# For static file serving
aiofiles==23.2.1